+++ CHANGELOG.md
@@ -13,8 +13,9 @@ Change Log
 ### Improvements
 
 - Select all matches by “Find All”.
-- Remove window size setting option.
+- Remove the window size setting option.
     - Now, the initial window size is determined by the last window size, just like other standard Cocoa applications.
+- Remove the initlal document inspector visibility setting option.
 - [trivial] Update editor opacity sample tip.